+++ dockci/models/job_meta/stages_prepare_docker.py
@@ -15,6 +15,7 @@ from dockci.exceptions import (AlreadyBuiltError,
                                StageFailedError,
                                )
 from dockci.models.base import ServiceBase
+from dockci.models.blob import FilesystemBlob
 from dockci.models.job_meta.stages import JobStageBase
 from dockci.util import (built_docker_image_id,
                          docker_ensure_image,
@@ -257,6 +258,58 @@ class ProvisionStage(InlineProjectStage):
         return True
 
 
+def parse_util_file(file_data):
+    """
+    Parse the input/output lines of util config
+
+    Examples:
+
+    >>> files = parse_util_file('/util /work/thing')
+    >>> files['from']
+    '/util'
+    >>> files['to']
+    '/work/thing'
+
+    >>> files = parse_util_file({'from': '/util', 'to': '/work/thing'})
+    >>> files['from']
+    '/util'
+    >>> files['to']
+    '/work/thing'
+
+    >>> files = parse_util_file('/work/thing')
+    >>> files['from']
+    '/work/thing'
+    >>> files['to']
+    '/work/thing'
+
+    >>> files = parse_util_file({'to': '/work/thing'})
+    >>> files['from']
+    '/work/thing'
+    >>> files['to']
+    '/work/thing'
+
+    >>> files = parse_util_file({'from': '/util'})
+    >>> files['from']
+    '/util'
+    >>> files['to']
+    '/util'
+    """
+    if isinstance(file_data, dict):
+        if 'from' not in file_data:
+            return {'from': file_data['to'], 'to': file_data['to']}
+        if 'to' not in file_data:
+            return {'from': file_data['from'], 'to': file_data['from']}
+
+        return file_data
+
+    else:
+        parts = file_data.split(' ')
+        if len(parts) == 1:
+            return {'from': file_data, 'to': file_data}
+        else:
+            return {'from': parts[0], 'to': parts[1]}
+
+
 class UtilStage(InlineProjectStage):
     """ Create, and run a utility stage container """
     def __init__(self, job, workdir, slug_suffix, config):
@@ -288,9 +341,13 @@ class UtilStage(InlineProjectStage):
           str: New image ID with files added
           bool: False if failure
         """
-        success = True
+        input_files = [
+            parse_util_file(input_data)
+            for input_data
+            in self.config.get('input', [])
+        ]
 
-        input_files = self.config.get('input', ())
+        success = True
         if not input_files:
             faux_log.update(progress="Skipped")
             return base_image_id
@@ -300,7 +357,10 @@ class UtilStage(InlineProjectStage):
         with tmp_file.open('w') as h_dockerfile:
             h_dockerfile.write('FROM %s\n' % base_image_id)
             for file_line in input_files:
-                h_dockerfile.write('ADD %s\n' % file_line)
+                h_dockerfile.write('ADD %s %s\n' % (
+                    file_line['from'],
+                    file_line['to'],
+                ))
 
         # Run the build
         rel_workdir = self.workdir.bestrelpath(tmp_file)
@@ -393,7 +453,11 @@ class UtilStage(InlineProjectStage):
         Returns:
           bool: True when all files retrieved as expected, False otherwise
         """
-        output_files = self.config.get('output', [])
+        output_files = [
+            parse_util_file(output_data)
+            for output_data
+            in self.config.get('output', [])
+        ]
         success = True
         if not output_files:
             faux_log.update(id=files_id, progress="Skipped")
@@ -509,20 +573,9 @@ class UtilStage(InlineProjectStage):
 
         return success
 
-    def runnable_inline(self, service, base_image_id, handle, faux_log):
+    def generate_data(self, service, base_image_id, handle, faux_log):
         """
-        Inline runner for utility projects. Adds files, runs the container,
-        retrieves output, and cleans up
-
-        Args:
-          service (dockci.models.base.ServiceBase): Service that this stage
-            uses the image from
-          base_image_id (str): Image ID of the utility base
-          handle: Stream handle for raw output
-          faux_log: The faux docker log object
-
-        Returns:
-          bool: True on all success, False on at least 1 failure
+        Adds files, runs the container, retrieves output, and cleans up
         """
         service_id = self.id_for_service(service.app_name)
 
@@ -588,6 +641,55 @@ class UtilStage(InlineProjectStage):
                                              )
 
         return success
+
+    def runnable_inline(self, service, base_image_id, handle, faux_log):
+        """
+        Inline runner for utility projects
+
+        Args:
+          service (dockci.models.base.ServiceBase): Service that this stage
+            uses the image from
+          base_image_id (str): Image ID of the utility base
+          handle: Stream handle for raw output
+          faux_log: The faux docker log object
+
+        Returns:
+          bool: True on all success, False on at least 1 failure
+        """
+        input_files = [parse_util_file(file_data)
+                       for file_data
+                       in self.config.get('input', ())]
+        output_files = [parse_util_file(file_data)
+                        for file_data
+                        in self.config.get('output', ())]
+        blob_store = None
+
+        if input_files:
+            blob_path = self.job.data_dir_path().join('_util_blobs')
+            blob_path.ensure_dir()
+            blob_store = FilesystemBlob.from_files(
+                blob_path,
+                self.workdir,
+                [
+                    self.workdir.join(input_data['from'])
+                    for input_data
+                    in input_files
+                ],
+                meta={'image': service.image},
+            )
+
+            for output_data in output_files:
+                blob_store.add_data(output_data['to'])
+
+        if blob_store and blob_store.exists:
+            blob_store.extract()
+            return True
+
+        else:
+            ret = self.generate_data(service, base_image_id, handle, faux_log)
+            if blob_store:
+                blob_store.write()
+            return ret
 
     @classmethod
     def slug_suffixes(cls, utility_names):